requirements.txt
pandas
sqlalchemy
requests
python-dotenv